Added initial release
Initial release with:
matlab simulation framework
python evaluation of algorithms
python network generation

project structure seems good enough for now.
still needs to be tested:
 - pip installing net_gen an eval
 - using net_gen and eval

do we need a top level pyproject.toml? eval and net-gen are never
used with eachother or with matlab code...

examples need to be polished.

This is version 0.0.1alpha, we still need to bring the newest
versions and add the more complicated examples.